requirements.txt
ccxt
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
mplfinance==0.12.9b7
multidict==6.0.4
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==20.2
pytz==2023.3
requests==2.29.0
rfc3986==1.5.0
schedule==1.2.0
six==1.16.0
sniffio==1.3.0
tzdata==2023.3
urllib3==1.26.15
yarl==1.9.2
zipp==3.15.0
